+++ containup/infra/runner/runner.py
@@ -4,6 +4,8 @@ import docker
 
 from containup import containup_cli, Config
 from containup.business.audit.audit_registry import AuditRegistry
+from containup.business.plugins.plugin_builtins import PluginBuiltins
+from containup.business.plugins.plugin_registry import PluginRegistry, register
 from containup.commands.command_down import CommandDown
 from containup.commands.command_up import CommandUp
 from containup.commands.execution_auditor import StdoutAuditor
@@ -28,6 +30,9 @@ class StackRunner:
         self.config = config or containup_cli()
         self.client: docker.DockerClient = docker.from_env()
         self._auditor = StdoutAuditor(self.stack, self.config)
+        register(PluginBuiltins)
+        self._plugin_registry = PluginRegistry()
+        self._audit_registry = AuditRegistry(self._plugin_registry)
         self.operator = (
             DryRunOperator(self._auditor)
             if self.config.dry_run
@@ -38,7 +43,7 @@ class StackRunner:
     # Handle command line parsing and launches the commands on the stack
     def run(self):
 
-        alerts = AuditRegistry().inspect(self.stack)
+        alerts = self._audit_registry.inspect(self.stack)
         if self.config.command == "up":
             CommandUp(self.stack, self.operator, self.user_interactions).up(
                 self.config.services